EmojiReactor/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.12.9",
    "discord-py>=2.5.2",
    "openai>=1.84.0",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
]
